Fix issue for adding 1st and subsequent properties
The most significant changes include the removal of a condition for displaying a div in `property-profiles.page.html`, the addition of a new method in `property-profiles.page.ts`, a change in navigation route in `property-profile-exterior-areas.page.ts`, and modifications in `user-types-manager.page.ts` to filter out certain user types and modify the redirection method.

1. In `property-profiles.page.html`, the condition for displaying the `home-btn-wrapper` div has been removed, causing the div to always be displayed regardless of the `isPrivateLabelUser` value.

2. A new method `dismissSpinner` has been added to the `PropertyProfilesPage` class in `property-profiles.page.ts`. This method sets `loadingVisible` to `false` and `spinnerText` to an empty string.

3. The navigation route after clicking "Yes" in `property-profile-exterior-areas.page.ts` has been changed from `property-profile-address` to `property-profile-general-information`.

4. In `user-types-manager.page.ts`, the `userTypes` array now filters out user types where the name is `PrivateLabelUser`.

5. The `redirectToRegistrationPage` method in `user-types-manager.page.ts` has been modified. The `debugger` statement has been removed and the `inAppBrowser` now creates a new page using the `PrivateLabelRegistrationUrl` instead of `x`.

Minor changes include the removal of a commented line of code and formatting changes in `property-profile-exterior-areas.page.ts`, and a non-functional change in the constructor of `PropertyProfileExteriorAreasPage` class in the same file.